enclosure/services/frontend/requirements.txt
flet>=0.22.0
paho-mqtt>=2.0.0
pyserial
